refactor: add base layout and restructure the dirs
