+++ cmd/web/main.go
@@ -2,14 +2,15 @@ package main
 
 import (
 	"fmt"
+	"myapp/pkg/handlers"
 	"net/http"
 )
 
 const portNumber = ":8080"
 
 func main() {
-	http.HandleFunc("/", Home)
-	http.HandleFunc("/about", About)
+	http.HandleFunc("/", handlers.Home)
+	http.HandleFunc("/about", handlers.About)
 
 	fmt.Printf("Application is running on %s", portNumber)
 	_ = http.ListenAndServe(portNumber, nil)
