--- handlers.go
@@ -1,24 +1,0 @@
-package main
-
-import (
-	"fmt"
-	"html/template"
-	"net/http"
-)
-
-func Home(w http.ResponseWriter, r *http.Request) {
-	renderTemplate(w, "/home.html")
-}
-
-func About(w http.ResponseWriter, r *http.Request) {
-	renderTemplate(w, "/about.html")
-}
-
-func renderTemplate(w http.ResponseWriter, tmpl string) {
-	parsedTemplate, _ := template.ParseFiles("./templates" + tmpl)
-	err := parsedTemplate.Execute(w, nil)
-	if err != nil {
-		fmt.Println("Error parsing the template", err)
-		return
-	}
-}
