+++ templates/home.html
@@ -1,12 +1,9 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Document</title>
-</head>
-<body>
-  <h1>Home Page</h1>
-</body>
-</html>
+{{template "base" .}} {{define "title"}} home {{end}} {{define "body"}}
+<div class="container">
+  <div class="row">
+    <div class="col">
+      <h1>Home Page</h1>
+    </div>
+  </div>
+</div>
+{{end}}
